lint: various
- short if / scoped if
- gofumpt
- unnecessary calls to Sprintf with const
- comment typo
- unnecessary type conversions
- replace deprecated `ioutil` calls
- struct alignment
These changes reduce the amount of pointer bytes scanned by the garbage
collector and reduces the size of 2 structs (`fidRef` & `connState`) by
 8 bytes each.

Considering that message types are frequently generated and short lived,
this should reduce pressure on the garbage collector even without
shrinking the actual size of the structs themselves.

Encoding and decoding are done via methods which are not dependent on
the memory layout, so this should not affect the wire format.