lint: struct alignment
These changes reduce the amount of pointer bytes scanned by the garbage
collector by a total of 520, and reduces the actual size for 2 structs
(`fidRef` & `connState`) by 8 each.

Considering that message types are frequently generated and short lived,
this should reduce pressure on the garbage collector even without
shrinking the actual size of the structs themselves.

Encoding and decoding are done via methods which are not dependent on
the memory layout, so this should not affect the wire format. Although
it may marginally affect encoding/decoding performance. The gains likely
outweigh the losses here in regard to CPU utilization, but this has not
been measured.